+++ src/pdfmm/base/PdfArray.h
@@ -14,6 +14,58 @@
 
 namespace mm {
 
+class PdfArray;
+
+class PdfArrayIterator
+{
+    friend class PdfArray;
+    typedef std::vector<PdfObject> Array;
+
+private:
+    PdfArrayIterator(PdfArray& arr);
+
+public:
+    template <typename TObject, typename TIterator>
+    class Iterator
+    {
+        friend class PdfArrayIterator;
+    public:
+        using difference_type = void;
+        using value_type = TObject;
+        using pointer = value_type*;
+        using reference = value_type&;
+        using iterator_category = std::forward_iterator_tag;
+    private:
+        Iterator(const TIterator& iterator);
+    public:
+        Iterator(const Iterator&) = default;
+        Iterator& operator=(const Iterator&) = default;
+        bool operator==(const Iterator& rhs) const;
+        bool operator!=(const Iterator& rhs) const;
+        Iterator& operator++();
+        reference operator*();
+        pointer operator->();
+    private:
+        void resolve();
+    private:
+        TIterator m_iterator;
+        bool m_resolved;
+        pointer m_ptr;
+    };
+
+    using iterator = Iterator<PdfObject, Array::iterator>;
+    using const_iterator = Iterator<const PdfObject, Array::const_iterator>;
+
+public:
+    iterator begin();
+    iterator end();
+    const_iterator begin() const;
+    const_iterator end() const;
+
+private:
+    PdfArray* m_arr;
+};
+
 /** This class represents a PdfArray
  *  Use it for all arrays that are written to a PDF file.
  *
@@ -89,6 +141,10 @@ public:
     void AddIndirect(const PdfObject& obj);
 
     void SetAtIndirect(const PdfObject& obj, unsigned idx);
+
+    PdfArrayIterator GetIndirectIterator();
+
+    const PdfArrayIterator GetIndirectIterator() const;
 
 public:
     /** Adds a PdfObject to the array
@@ -250,6 +306,67 @@ void PdfArray::insert(const PdfArray::iterator& pos,
     }
 
     SetDirty();
+}
+
+template<typename TObject, typename TIterator>
+PdfArrayIterator::Iterator<TObject, TIterator>::Iterator(const TIterator& iterator)
+    : m_iterator(iterator), m_resolved(false), m_ptr(nullptr) { }
+
+template<typename TObject, typename TIterator>
+inline bool PdfArrayIterator::Iterator<TObject, TIterator>::operator==(const Iterator& rhs) const
+{
+    return m_iterator == rhs.m_iterator;
+}
+
+template<typename TObject, typename TIterator>
+bool PdfArrayIterator::Iterator<TObject, TIterator>::operator!=(const Iterator& rhs) const
+{
+    return m_iterator != rhs.m_iterator;
+}
+
+template<typename TObject, typename TIterator>
+PdfArrayIterator::Iterator<TObject, TIterator>& PdfArrayIterator::Iterator<TObject, TIterator>::operator++()
+{
+    m_iterator++;
+    m_resolved = false;
+    return *this;
+}
+
+template<typename TObject, typename TIterator>
+typename PdfArrayIterator::Iterator<TObject, TIterator>::reference PdfArrayIterator::Iterator<TObject, TIterator>::operator*()
+{
+    resolve();
+    return *m_ptr;
+}
+
+template<typename TObject, typename TIterator>
+typename PdfArrayIterator::Iterator<TObject, TIterator>::pointer PdfArrayIterator::Iterator<TObject, TIterator>::operator->()
+{
+    resolve();
+    return m_ptr;
+}
+
+template<typename TObject, typename TIterator>
+inline void PdfArrayIterator::Iterator<TObject, TIterator>::resolve()
+{
+    if (m_resolved)
+        return;
+
+    TObject& robj = *m_iterator;
+    TObject* indirectobj;
+    PdfReference ref;
+    if (robj.TryGetReference(ref)
+        && ref.IsIndirect()
+        && (indirectobj = robj.GetDocument()->GetObjects().GetObject(ref)) != nullptr)
+    {
+        m_ptr = indirectobj;
+    }
+    else
+    {
+        m_ptr = &robj;
+    }
+
+    m_resolved = true;
 }
 
 };